Update FPA422Handler.cpp
Had to reshuffle  the Wifi connections: First try to connect to the known external network. If that fails, determine the best WiFi channel and have the access point use this channel.
If you first set-up the access point, and then try to connect to the external Wifi network and this fails, you cannot connect to the access point anymore.